move multi_rotate_signers to tests/auth.rs
