+++ contracts/axelar-gateway/src/auth.rs
@@ -235,56 +235,3 @@ fn validate_signers(env: &Env, weighted_signers: &WeightedSigners) -> Result<(),
 
     Ok(())
 }
-
-#[cfg(all(test, feature = "testutils"))]
-mod tests {
-    use soroban_sdk::{testutils::BytesN as _, BytesN, Env};
-
-    use crate::auth;
-    use axelar_soroban_std::assert_ok;
-
-    use crate::testutils::{self, generate_proof, generate_signers_set, randint, setup_gateway};
-
-    #[test]
-    fn multi_rotate_signers() {
-        let previous_signer_retention = randint(1, 5);
-        let env = Env::default();
-        env.mock_all_auths();
-        let (original_signers, client) =
-            setup_gateway(&env, previous_signer_retention, randint(1, 10));
-
-        let msg_hash: BytesN<32> = BytesN::random(&env);
-
-        let mut previous_signers = original_signers.clone();
-
-        for _ in 0..previous_signer_retention {
-            let new_signers = generate_signers_set(
-                &env,
-                randint(1, 10),
-                original_signers.domain_separator.clone(),
-            );
-
-            testutils::rotate_signers(&env, &client.address, new_signers.clone());
-
-            let proof = generate_proof(&env, msg_hash.clone(), new_signers.clone());
-
-            env.as_contract(&client.address, || {
-                assert!(assert_ok!(auth::validate_proof(&env, &msg_hash, proof)));
-            });
-
-            let proof = generate_proof(&env, msg_hash.clone(), previous_signers.clone());
-
-            env.as_contract(&client.address, || {
-                assert!(!assert_ok!(auth::validate_proof(&env, &msg_hash, proof)));
-            });
-
-            previous_signers = new_signers;
-        }
-
-        // Proof from the first signer set should still be valid
-        let proof = generate_proof(&env, msg_hash.clone(), original_signers.clone());
-        env.as_contract(&client.address, || {
-            assert!(!assert_ok!(auth::validate_proof(&env, &msg_hash, proof)));
-        })
-    }
-}
